refactor to FastAPI
Moving forward, it is recommended to switch to FastAPI as soon as
possible from Flask. Better performance, concurrency, and scalability as
the project grows. We can keep the frontend simple, but I envision the
server side logic quickly growing in complexity as we implement
integrations and business logic - especially looking at the immediate
scope of serving as a recommendations/discovery engine. algorithms and
shit. let's get this bread